% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.bclogit.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Helper to get compiled stan model from cache}
\usage{
get_stan_model(file_name)
}
\description{
Helper to get compiled stan model from cache
}
\keyword{internal}
