% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bclogit.R
\name{print.summary.bclogit}
\alias{print.summary.bclogit}
\title{Print summary of a bclogit model}
\usage{
\method{print}{summary.bclogit}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A \code{summary.bclogit} object.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print summary of a bclogit model
}
\examples{
\donttest{
data("fhs")
fit <- bclogit(PREVHYP ~ TOTCHOL + BMI, data = fhs,
               treatment = PERIOD, strata = RANDID)
print(summary(fit))
}
}
\seealso{
\code{\link{summary.bclogit}}, \code{\link{confint.bclogit}}
}
