% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.bclogit.R
\name{vcov.bclogit}
\alias{vcov.bclogit}
\title{Extract variance-covariance matrix from a bclogit model}
\usage{
\method{vcov}{bclogit}(object, ...)
}
\arguments{
\item{object}{A \code{bclogit} object.}

\item{...}{Additional arguments.}
}
\value{
A matrix of the estimated covariance of the coefficients.
}
\description{
Extract variance-covariance matrix from a bclogit model
}
