% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.clogit_bclogit.R
\name{vcov.clogit_bclogit}
\alias{vcov.clogit_bclogit}
\title{Extract variance-covariance matrix from a clogit_bclogit model}
\usage{
\method{vcov}{clogit_bclogit}(object, ...)
}
\arguments{
\item{object}{A \code{clogit_bclogit} object.}

\item{...}{Additional arguments.}
}
\value{
A matrix of the estimated covariance of the coefficients.
}
\description{
Extract variance-covariance matrix from a clogit_bclogit model
}
