% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclogit.R
\docType{package}
\name{bclogit-package}
\alias{bclogit-package}
\title{Bayesian Conditional Logistic Regression with Concordant Pairs}
\description{
bclogit
}
\details{
Fits a conditional logistic regression model to incidence data. Allows for use of the concordant pairs in the fitting.
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{paper to go here
}\if{html}{\out{</div>}}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	Jacob Tennenbaum \email{Jacob.Tennenbaum51@qmail.cuny.edu}
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{	Adam Kapelner \email{kapelner@qc.cuny.edu}
}\if{html}{\out{</div>}}
}
\keyword{Conditional}
\keyword{Logistic}
\keyword{Regression}
