% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclogit.R
\docType{package}
\name{bclogit-package}
\alias{bclogit-package}
\title{Bayesian Conditional Logistic Regression with Concordant Pairs}
\description{
bclogit
}
\details{
Fits a conditional logistic regression model to incidence data. Allows for use of the concordant pairs in the fitting.
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Jacob Tennenbaum and Adam Kapelner (2026). "Improved Conditional Logistic Regression using Information in Concordant Pairs with Software." arXiv preprint arXiv:2602.08212.
}\if{html}{\out{</div>}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tennenbaumj/bclogit_package_and_paper_repo}
  \item \url{https://tennenbaumj.github.io/bclogit_package_and_paper_repo/}
  \item Report bugs at \url{https://github.com/tennenbaumj/bclogit_package_and_paper_repo/issues}
}

}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	Jacob Tennenbaum \email{Jacob.Tennenbaum51@qmail.cuny.edu}
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{	Adam Kapelner \email{kapelner@qc.cuny.edu}
}\if{html}{\out{</div>}}
}
\keyword{Conditional}
\keyword{Logistic}
\keyword{Regression}
