% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclogit.R, R/bclogitFunctions.R,
%   R/bclogitMethods.R
\docType{package}
\name{bclogit}
\alias{bclogit-package}
\alias{bclogit}
\alias{bclogit.default}
\alias{bclogit.formula}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{	Conditional Logistic Regression with Optional Concordant pairs
}\if{html}{\out{</div>}}}
\usage{
bclogit(x, ...)

\method{bclogit}{default}(
  response = NULL,
  data = NULL,
  treatment = NULL,
  strata = NULL,
  concordant_method = "GEE",
  prior_type = "hybrid",
  treatment_name = NULL
)

\method{bclogit}{formula}(formula, data, treatment = NULL, strata = NULL, ...)
}
\arguments{
\item{response}{A binary (0,1) vector containing the response of each subject.}

\item{data}{A data.frame, data.table, or model.matrix containing the variables.}

\item{treatment}{A binary (0,1) vector containing whether each subject is treatment or control.}

\item{strata}{Numeric vector indexing the matched pairs and the reservoir. Entries indexed with zero, are added to the reservoir. Entries indexed with the same number are part of the same strata.}

\item{concordant_method}{The method to use for fitting the concordant pairs and reservoir. Options are "GLM", "GEE", and "GLMM".}

\item{prior_type}{The type of prior to use for the discordant pairs. Options are "naive", "G prior", "PMP", and "hybrid".}

\item{treatment_name}{Optional string name for the treatment variable.}
}
\value{
A list of class \code{bclogit}.
}
\description{
bclogit
}
\details{
Fits a conditional logistic regression model to incidence data. Allows for use of the concordant pairs in the fitting.
}
\section{Methods (by class)}{
\itemize{
\item \code{bclogit(default)}: Default method

\item \code{bclogit(formula)}: Formula method

}}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{paper to go here
}\if{html}{\out{</div>}}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	Jacob Tennenbaum \email{Jacob.Tennenbaum51@qmail.cuny.edu}
}\if{html}{\out{</div>}}
}
\keyword{Conditional}
\keyword{Logistic}
\keyword{Regression}
