% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclogitMethods.R
\name{coef.bclogit}
\alias{coef.bclogit}
\title{Extract coefficients from a bclogit model}
\usage{
\method{coef}{bclogit}(object, ...)
}
\arguments{
\item{object}{A \code{bclogit} object.}

\item{...}{Additional arguments.}
}
\value{
Numeric vector of coefficients.
}
\description{
Extract coefficients from a bclogit model
}
