% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclogitMethods.R
\name{confint.bclogit}
\alias{confint.bclogit}
\title{Credible Intervals for bclogit Parameters}
\usage{
\method{confint}{bclogit}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{bclogit} object.}

\item{parm}{A specification of which parameters to be given credible intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required (default 0.95).}

\item{...}{Additional arguments.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter.
}
\description{
Computes Bayesian credible intervals for the model parameters.
}
