% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.bclogit.R
\name{formula.bclogit}
\alias{formula.bclogit}
\title{Extract model formula}
\usage{
\method{formula}{bclogit}(x, ...)
}
\arguments{
\item{x}{A \code{bclogit} object.}

\item{...}{Additional arguments.}
}
\value{
The formula used in the model.
}
\description{
Extract model formula
}
