% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bclogit.R
\name{print.summary.bclogit}
\alias{print.summary.bclogit}
\title{Print summary of a bclogit model}
\usage{
\method{print}{summary.bclogit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.bclogit} object.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional arguments.}
}
\description{
Print summary of a bclogit model
}
