% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clogit_bclogit.R
\name{summary.clogit_bclogit}
\alias{summary.clogit_bclogit}
\title{Summarize a clogit_bclogit model}
\usage{
\method{summary}{clogit_bclogit}(object, ...)
}
\arguments{
\item{object}{A \code{clogit_bclogit} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{summary.clogit_bclogit} object.
}
\description{
Summarize a clogit_bclogit model
}
